% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_comma.R
\name{add_comma}
\alias{add_comma}
\title{Label numbers in decimal format (e.g. 1,234)}
\usage{
add_comma(data, col)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble),
or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{col}{Columns that inserts a comma every three digits}
}
\description{
\code{add_comma()} is a function that inserts a comma every three digits for columns in a data frame.
}
\examples{
df <- add_comma(covid_raw, c("total_cases", "new_cases"))
view(df)

}
