% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_about.R
\name{covid_about}
\alias{covid_about}
\title{Add about page into shiny app}
\usage{

}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble),
or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{col}{Columns that inserts a comma every three digits}
}
\description{
\code{add_comma()} is a function that inserts a comma every three digits for columns in a dataframe.
}
\details{

}
\examples{
df <- add_comma(covid_raw, c("total_cases", "new_cases"))
view(df)

}
