% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_raw.R
\docType{data}
\name{covid_raw}
\alias{covid_raw}
\title{World wide COVID-19 Pandemic dataset}
\format{
Data frame with columns
\describe{
\item{iso_code, continent, location}{Location information}
\item{date}{Date in YYYY-MM-DD format.}
\item{total_cases, new_cases, new_cases_smoothed, total_cases_per_million, new_cases_per_million, new_cases_smoothed_per_million}{Case-related numbers}
\item{total_deaths, new_deaths, new_deaths_smoothed, total_deaths_per_million, new_deaths_per_million, new_deaths_smoothed_per_million}{Death-related numbers}
\item{new_tests, total_tests, total_tests_per_thousand, new_tests_per_thousand, new_tests_smoothed, new_tests_smoothed_per_thousand, tests_per_case, positive_rate, tests_units}{Test-related numbers}
}
}
\source{
Our World in Data
\url{https://ourworldindata.org/coronavirus}
}
\usage{
covid_raw
}
\description{
This dataset contains COVID-19 information of 212 countries from December 31, 2019, to September 27, 2020.
}
\details{
The data contains the daily summary of COVID-19 cases
(confirmed and death), by country.
}
\examples{
library(COVID19dashboard)
covid_raw



}
\keyword{datasets}
