% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_range.R
\name{date_range}
\alias{date_range}
\title{Create date range input}
\usage{
date_range(data, date_col = "date", input_id)
}
\arguments{
\item{data}{The data used to create tables ot plots which are correlated to the input}

\item{date_col}{The column name in the data which contain the date information. The default value is \code{date}.}

\item{input_id}{The input slot that will be used to access the value.}
}
\description{
\code{date_range} is a function that create date range input into shiny app.
When clicked on, bring up calendars that the user can click on to select dates.
}
\details{
The date \code{format} string specifies how the date will be displayed in the browser. It allows the following values:
\itemize{
\item \code{yy} Year without century (12)
\item \code{yyyy} Year with century (2012)
\item \code{mm} Month number, with leading zero (01-12)
\item \code{m} Month number, without leading zero (1-12)
\item \code{M} Abbreviated month name
\item \code{MM} Full month name
\item \code{dd} Day of month with leading zero
\item \code{d} Day of month without leading zero
\item \code{D} Abbreviated weekday name
\item \code{DD} Full weekday name
}
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  date_range(covid_raw, "daterange"))
                 
shinyApp(ui, server = function(input, output) { })
}

}
\seealso{
\code{\link[=dateInput]{dateInput()}}, \code{\link[=updateDateRangeInput]{updateDateRangeInput()}}
}
