% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the COVID-19 shiny dashboard}
\arguments{
\item{data}{A dataset containing COVID-19 relevant information.}
}
\description{
The \code{launch_app()} function is used to load the coronavirus pandemic shiny application.
}
\details{
The coronavirus pandemic, also known as the COVID-19 pandemic.
The disease was first identified in December 2019 in Wuhan, China.
The outbreak was declared a Public Health Emergency of International Concern in January 2020, and a pandemic in March 2020.
As of 11 October 2020, more than 37.2 million cases have been confirmed as well as more than 1.07 million deaths attributed to COVID-19.
The aim of this shiny dashboard is to complement the raw data by providing interactive visualisation which is not available by the news, including the timeline function and the ability to compare between countries.
}
