% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_note.R
\name{shiny_note}
\alias{shiny_note}
\title{Add a note box into shiny app}
\usage{
shiny_note(...)
}
\arguments{
\item{...}{HTML tags used to add text into shiny app}
}
\description{
\code{shiny_note()} is a function that inserts a box that contains elements of note.
Using HTML tags to write you own notes for your shiny app,
including How to interactive with the app, what does the output mean
}
\examples{
\dontrun{
shiny_note(
    tags$h3("Background & Purpose of this app"),
    h5("In December 2019..."),
    tags$h3("Code"),
    "Code and input data used to generate this Shiny app are available on..."
    )
}

}
