% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_raw.R
\docType{data}
\name{covid_raw}
\alias{covid_raw}
\title{World wide COVID-19 Pandemic dataset}
\format{
Data frame with columns
\describe{
\item{iso_code, continent, location}{Location information}
\item{date}{Recording date}
}
}
\source{
Our World in Data
\url{https://ourworldindata.org/coronavirus}
}
\usage{
covid_raw
}
\description{
This dataset contains COVID-19 information of 212 countries from December 31, 2019 to September 27, 2020.
}
\examples{
library(COVID19dashboard)
covid_raw



}
\keyword{datasets}
